+++ src/main/java/com/rojas/dev/XCampo/entity/CartItem.java
@@ -21,8 +21,8 @@ public class CartItem {
     private Product product;
 
     @Column(nullable = false)
-    private int quantity;
+    private Long quantity;
 
     @Column(nullable = false)
-    private double unitPrice;
+    private Long unitPrice;
 }
